ADDED: Database seed, More piles visuals, backend links for piles
Kayla and I met today to combine our ideas to connect piles from  back end to the front end. This includes showing all the created piles on the front end. Showing better visuals of the piles.

TODO: We want to have the visuals of the cards on top of the piles to be an option, to do so, we need to look further into img tags to change the background images of the pile class.
Co-Authored-By: Kayla Felderman <kmfelderman@uiowa.edu>