TEST: Testing Visibility Function within piles
We have tested the Visibility function within the piles. Currently, the only way to change the visibilty of the piles is to go into the Dev Database and change the visible object variable. Refresh page, and then the page will show only the back of the card or the King of Hearts.

Co-Authored-By: Kayla Felderman <kmfelderman@uiowa.edu>