ENH: Card images
Got card images from asset pipeline to display correctly. This enables us to attach a string to each card object in the database containing the filename of the image associated with that card. This is then displayed on the screen so that the user can see which cards they have, or are in play.