TEST: Dropping DB to add new column
Adding functionality to the buttons once again while fixing the layer. ERROR: When having cards visible, somehow the divs get messed up to where it moves piles around on the website.

ADDITIONALLY: I will be adding a new column to the database called image or png to signify the image attached to the card.